+++ Data/SBV/Examples/BitPrecise/PrefixSum.hs
@@ -141,8 +141,8 @@ thm2 = prove $ flIsCorrect 16 (0, smax)
 --   s18
 ladnerFischerTrace :: Int -> IO ()
 ladnerFischerTrace n = gen >>= print
-  where gen = runSymbolic (True, Nothing) $ do args :: [SWord8] <- mkForallVars n
-                                               mapM_ output $ lf (0, (+)) args
+  where gen = runSymbolic (True, defaultSMTCfg) $ do args :: [SWord8] <- mkForallVars n
+                                                     mapM_ output $ lf (0, (+)) args
 
 -- | Trace generator for the reference spec. It clearly demonstrates that the reference
 -- implementation fewer operations, but is not parallelizable at all:
@@ -186,7 +186,7 @@ ladnerFischerTrace n = gen >>= print
 --
 scanlTrace :: Int -> IO ()
 scanlTrace n = gen >>= print
-  where gen = runSymbolic (True, Nothing) $ do args :: [SWord8] <- mkForallVars n
-                                               mapM_ output $ ps (0, (+)) args
+  where gen = runSymbolic (True, defaultSMTCfg) $ do args :: [SWord8] <- mkForallVars n
+                                                     mapM_ output $ ps (0, (+)) args
 
 {-# ANN module ("HLint: ignore Reduce duplication" :: String) #-}